Move macro-related stuff to separate file
